Added I2C1 library, changed BME280 to use this.
