--- EmbeddedSoftMainProgram/EmbeddedSoftMainProgram/.gitignore
@@ -1,57 +1,0 @@
-# Created by https://www.toptal.com/developers/gitignore/api/keil
-# Edit at https://www.toptal.com/developers/gitignore?templates=keil
-
-### Keil ###
-# something maybe need to add
-# *.bat  # generate batch file
-# output/project_name # uv project generate without file ext
-# *.SRC # generate source file
-# *.src
-
-# Prerequisites
-*.d
-*.dep
-# Object files
-*.o
-*.obj
-*.OBJ
-
-# Listing Files
-*.COD
-*.cod
-*.[Ii]
-*.LST
-*.lst
-*.MAP
-*.map
-*.[Mm]51
-*.[Mm]66
-*.SCR
-*.scr
-
-# LINK output
-*.ABS
-
-# Libraries
-*.lib
-*.LIB
-
-
-# uv output
-*.lnp
-*.build_log.htm
-*.HEX
-*.hex
-
-
-# Build Files
-*._IA
-*.__[Ii]
-*._[Ii][Ii]
-
-# project
-*.uvgui.*
-*.uvguix.*
-
-
-# End of https://www.toptal.com/developers/gitignore/api/keil
