--- EmbeddedSoftMainProgram/EmbeddedSoftMainProgram/RTE/_Target_1/RTE_Components.h
@@ -1,21 +1,0 @@
-
-/*
- * Auto generated Run-Time-Environment Configuration File
- *      *** Do not modify ! ***
- *
- * Project: 'TimersLab' 
- * Target:  'Target 1' 
- */
-
-#ifndef RTE_COMPONENTS_H
-#define RTE_COMPONENTS_H
-
-
-/*
- * Define the Device Header File: 
- */
-#define CMSIS_device_header "MKL25Z4.h"
-
-
-
-#endif /* RTE_COMPONENTS_H */
