--- EmbeddedSoftMainProgram/EmbeddedSoftMainProgram/uart0/uart0.h
@@ -1,45 +1,0 @@
-/*! ***************************************************************************
- *
- * \brief     Uart driver
- * \file      uart0.h
- * \author    Hugo Arends
- * \date      July 2021
- *
- * \copyright 2021 HAN University of Applied Sciences. All Rights Reserved.
- *            \n\n
- *            Permission is hereby granted, free of charge, to any person
- *            obtaining a copy of this software and associated documentation
- *            files (the "Software"), to deal in the Software without
- *            restriction, including without limitation the rights to use,
- *            copy, modify, merge, publish, distribute, sublicense, and/or sell
- *            copies of the Software, and to permit persons to whom the
- *            Software is furnished to do so, subject to the following
- *            conditions:
- *            \n\n
- *            The above copyright notice and this permission notice shall be
- *            included in all copies or substantial portions of the Software.
- *            \n\n
- *            THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND,
- *            EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES
- *            OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND
- *            NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT
- *            HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY,
- *            WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING
- *            FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR
- *            OTHER DEALINGS IN THE SOFTWARE.
- *
- *****************************************************************************/
-#ifndef UART0_H
-#define UART0_H
-
-#include <stdint.h>
-#include <MKL25Z4.h>
-#include "queue.h"
-
-void uart0_init(void);
-uint32_t uart0_num_rx_chars_available(void);
-char uart0_get_char(void);
-void uart0_put_char(char c);
-void uart0_send_string(char *str);
-
-#endif
