+++ EmbeddedSoftMainProgram/EventRecorderStub.scvd
